Split the interview page script into the modules it grew
web/interview.js was 2235 lines and 85 top-level functions. Thirteen
sibling modules had already been extracted from it, so what was left was
the wiring plus six clusters that never left: the avatar and its analyser,
the integrity timers and worker, the replay feed, the recording lifecycle,
the caption reveal, and audio output routing. Each is a set of timers or a
queue with one lifetime, and what they guard against are teardown-order
bugs, which are the kind that read as fine in a file where everything else
is also in scope.

Six modules now, and the page script is 1474 lines. Every moved line is the
one it was: each module takes the page's bindings through an init call into
names it already used, so nothing had to be rewritten to reach state and no
assertion had to be loosened to keep matching. Where a module needs
something the page owns, it takes a function rather than importing back, so
nothing is circular. The presence table goes the other way, into the module
that reads it, because it is the vocabulary of the worker's messages and
touches neither the DOM nor the interview.

Two things stopped being reachable from outside the file that owns them.
The avatar's renderer and analyser track were read from three places in the
page; those are accessor functions now, so the renderer has one owner and
the teardown order is not something a caller can get wrong. The replay
throttles were inline guards at their call sites reaching into replay
state, and they move to the module that owns the interval.

The source-text tests needed care rather than re-pointing. Assertions like
"setSinkId targets Jim and nothing else" are properties of a path, not of a
filename, and a per-file scan silently stops covering the half that moved,
which is the fail-open tests/browser/source.js was written to warn about.
They read one list of the interview path instead, and a test derives that
list back out of the source, so a seventh module cannot join the tree
without joining the list.

One specifier did not survive the move and nothing caught it. The renderer
is imported lazily, and the path that was correct from the page script
named a file one directory too deep from web/avatar/. The assertion meant
to pin it matched the literal string either way, and the import only runs
once a published model answers a HEAD. It is fixed, the assertion now names
the resolved path, and every relative specifier in every first-party script
is checked against the filesystem.
