+++ tests/browser/recording-consent.test.js
@@ -13,10 +13,13 @@
 
 import { test } from "node:test";
 import assert from "node:assert/strict";
-import { functionBody, read } from "./source.js";
+import { functionBody, interviewSource, read } from "./source.js";
 
 const page = read("web/interview.html");
-const script = read("web/interview.js");
+// The consent path spans the page script and the replay queue it starts
+// and stops, so both read as one text: an assertion that consent closes
+// the feed must not be satisfiable by the feed moving to a sibling file.
+const script = interviewSource();
 
 /// Source with line comments removed.
 ///
@@ -159,6 +162,21 @@ test("recording-consent is recorded before a token is asked for", () => {
     /interviewId/.test(body.slice(token)),
     "and the token request carries the interview the consent belongs to",
   );
+});
+
+
+/// The poll outlives the room otherwise: a request every fifteen seconds about
+/// an interview that ended, asking for a word that will not change again.
+test("the recording poll is stopped, not just abandoned", () => {
+  const stop = withoutComments(functionBody(script, "stopRecordingPoll"));
+  assert.ok(stop.includes("clearInterval(recordingPoll);"), "the interval must be cleared");
+  assert.ok(stop.includes("recordingPoll = null;"), "and the handle dropped so a restart cannot double it");
+  for (const caller of ["pollRecordingState", "showRecordingState"]) {
+    assert.ok(
+      withoutComments(functionBody(script, caller)).includes("stopRecordingPoll()"),
+      `${caller} must stop the poll rather than leave it running`,
+    );
+  }
 });
 
 test("recording-consent sends the version it displayed", () => {
